pyproject.toml
[project]
name = "echoforge"
version = "0.1.0"
description = "Simple chatbot powered by LangChain + OpenRouter (Gemini)"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.0.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.4.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]
